+++ pytorch_lightning/metrics/functional/ir_hit_rate.py
@@ -14,29 +14,31 @@
 import torch
 
 
-def average_precision(
+def retrieval_hit_rate(
     preds: torch.Tensor,
-    target: torch.Tensor
+    target: torch.Tensor,
+    k: int = 1
 ):
     """
-    Computes average precision metric for information retrieval,
-    as explained here: https://en.wikipedia.org/wiki/Evaluation_measures_(information_retrieval)#Mean_average_precision
+    Computes the hit_rate @ k metric for information retrieval
+    Hir Rate at k is 1 iff there is at least one relevant documents among the top K.
 
     `preds` and `target` should be of the same shape and live on the same device. If not target is true, 0 is returned.
 
     Args:
         preds: estimated probabilities of each document to be relevant.
         target: ground truth about each document being relevant.
-    
+        k: consider only the top k elements.
+
     Returns:
-        a single-value tensor with the average precision (AP) of the predictions `preds` wrt the labels `target`.
+        a single-value tensor with the hit rate at k (HR@K) of the predictions `preds` wrt the labels `target`.
 
     Example:
 
         >>> preds = torch.tensor([0.2, 0.3, 0.5])
         >>> target = torch.tensor([True, False, True])
-        >>> average_precision(preds, target)
-        ... 0.833
+        >>> hit_rate(preds, target, k=2)
+        ... 1.0
     """
 
     if preds.shape != target.shape or preds.device != target.device: 
@@ -44,10 +46,5 @@ def average_precision(
             f"`preds` and `target` must have the same shape and be on the same device"
         )
 
-    if target.sum() == 0:
-        return torch.tensor(0).to(preds)
-
-    target = target[torch.argsort(preds, dim=-1, descending=True)]
-    positions = torch.arange(1, len(target) + 1, device=target.device)[target > 0]
-    res = torch.true_divide((torch.arange(len(positions), device=positions.device) + 1), positions).mean()
-    return res
+    relevant = target[torch.argsort(preds, dim=-1, descending=True)][:k].sum()
+    return (relevant > 0).to(preds)
