+++ pytorch_lightning/metrics/retrieval/__init__.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #from pytorch_lightning.metrics.retrieval import 
-from pytorch_lightning.metrics.retrieval.mean_average_precision import MeanAveragePrecision
-from pytorch_lightning.metrics.retrieval.mean_reciprocal_rank import MeanReciprocalRank
+from pytorch_lightning.metrics.retrieval.mean_average_precision import RetrievalMAP
+from pytorch_lightning.metrics.retrieval.mean_reciprocal_rank import RetrievalMRR
 from pytorch_lightning.metrics.retrieval.precision import RetrievalPrecision
 from pytorch_lightning.metrics.retrieval.recall import RetrievalRecall
-from pytorch_lightning.metrics.retrieval.hit_rate import HitRate
+from pytorch_lightning.metrics.retrieval.hit_rate import RetrievalHitRate
