+++ pytorch_lightning/metrics/retrieval/mean_average_precision.py
@@ -1,10 +1,10 @@
 from typing import List
 
 from pytorch_lightning.metrics.retrieval.retrieval_metric import RetrievalMetric
-from pytorch_lightning.metrics.functional.average_precision import average_precision
+from pytorch_lightning.metrics.functional.ir_average_precision import retrieval_average_precision
 
 
-class MeanAveragePrecision(RetrievalMetric):
+class RetrievalMAP(RetrievalMetric):
     """
     Mean Average Precision computes the MAP over multiple retrieved documents for each query.
     Each average precision computation on a single query can be done on a different number of 
@@ -16,7 +16,7 @@ class MeanAveragePrecision(RetrievalMetric):
         >>> preds = torch.tensor([0.2, 0.3, 0.5, 0.1, 0.3, 0.5, 0.2])
         >>> target = torch.tensor([False, False, True, False, True, False, False])
 
-        >>> map = MeanAveragePrecision()
+        >>> map = RetrievalMAP()
         >>> map(indexes, preds, target)
         >>> map.compute()
         ... 0.75
@@ -26,4 +26,4 @@ class MeanAveragePrecision(RetrievalMetric):
         _preds = self.preds[group]
         _target = self.target[group]
         valid_indexes = _target != self.exclude
-        return average_precision(_preds[valid_indexes], _target[valid_indexes])
+        return retrieval_average_precision(_preds[valid_indexes], _target[valid_indexes])
